requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appdirs==1.4.3
beautifulsoup4==4.6.0
boto3==1.7.31
botocore==1.10.32         # via boto3, s3transfer
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
contextlib2==0.5.4
directory-client-core==0.2.1  # via directory-sso-api-client
directory-constants==11.4.0  # via directory-components
directory-sso-api-client==2.5.1
directory-validators==5.0.0
directory_components==3.6.1
dj-database-url==0.4.1
django-admin-ip-restrictor==1.0.0  # via directory-components
django-appconf==1.0.1
django-ckeditor==5.0.3
django-environ==0.4.5
django-grappelli==2.8.3
django-image-cropping==1.0.4
django-ipware==2.1.0      # via django-admin-ip-restrictor
django-reversion==2.0.8
django-storages==1.5.2
django-thumber==1.0.0
django==1.11.17
djangorestframework==3.8.2  # via sigauth
docker-py==1.2.1
docutils==0.13.1
easy-thumbnails==2.5
export-elements==1.0.0    # via directory-components
gunicorn==19.6.0
idna==2.6                 # via requests
jmespath==0.9.1
markdown2==2.3.5
mohawk==0.3.4             # via directory-components, sigauth
monotonic==1.5            # via directory-client-core
olefile==0.44             # via directory-validators, pillow
packaging==16.8
phonenumberslite==7.7.4   # via directory-validators
pillow==4.2.1
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2              # via directory-validators, django
raven==5.31.0
requests==2.18.4
s3transfer==0.1.10
sigauth==2.1.2
six==1.10.0
urllib3==1.22             # via requests
whitenoise==3.2.1
whoosh==2.7.4
